Fix test
Exclude deleting the admin database from tests.

+++ examples/grails3-mongodb-database-per-tenant/src/integration-test/groovy/functional/tests/BookSpec.groovy
@@ -21,7 +21,9 @@ class BookSpec extends Specification {
     void "Test database per tenant"() {
         setup:
         mongoDatastore.mongoClient.listDatabaseNames().forEach( { String name ->
-            mongoDatastore.mongoClient.getDatabase(name).drop()
+            if (name != 'admin') {
+                mongoDatastore.mongoClient.getDatabase(name).drop()
+            }
         } as Block)
 
         when:"A query is executed"
